2nd commit

+++ src/main/java/tn/test/spring/Services/SecteurActivite/SecteurActiviteServiceImp.java
@@ -5,7 +5,7 @@ import tn.test.spring.Entity.SecteurActivite;
 import tn.test.spring.Services.GServiceImp;
 
 
-public class SecteurActiviteServiceImp extends GServiceImp<SecteurActivite, Integer> implements SecteurActiviteService {
+public class SecteurActiviteServiceImp extends GServiceImp<SecteurActivite, Long> implements SecteurActiviteService {
 
 
 }
